test/LayoutTests/Panel/clickMessages.html

<html>
<body>
<h1>Feature Test Click Messages issue #27</h1>
</body>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/ChannelPlate/ChannelPlate.js"></script>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/ChannelPlate/RemoteMethodCall.js"></script>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/AsyncMachine.js"></script>
<script src="../PatientSelector.js"></script>
<script src="../panel-test.js"></script>

<script>

window.addEventListener('load', function() {
  testScript();
});

function test_clickMessage() {
  console.log("------------test begins------------");
  AsyncMachine.pushOp(evaluateInPage, 'window.location="http://localhost:8686/test/qpd.html"');
  AsyncMachine.pushOp(openQuerypointPanel);
  AsyncMachine.pushOp(openSourceFile, 'demo.js');
  
  var objPropSelector = [
    {type: 'string', text:'\"I\'ve been clicked \"'},
    {type: 'property', text: 'prop'}
    ];
  AsyncMachine.pushOp(selectTokenInSource, objPropSelector);
  AsyncMachine.pushOp(verifyTokenView, 'obj.prop');
  AsyncMachine.pushOp(clickQPOperation, 'lastChange');
  AsyncMachine.pushOp(whenSelectorAll, '.currentLoadNumber', '1');

  for(var i=0;i<30;i++) AsyncMachine.pushOp(evaluateInPage, 'document.querySelector("#myButton").dispatchEvent(new MouseEvent("click"))', function(text){} );
  AsyncMachine.pushOp(whenSelectorAll, 'span.turnNumber', '↱31');
  AsyncMachine.pushOp(evaluateInPanel, 'document.querySelector("div.eventIndicator[severity=error][turn=\'10\']").dispatchEvent(new MouseEvent("mousedown"))', function(text){ return "";} );
  AsyncMachine.pushOp(evaluateInPanel, '((document.querySelectorAll("td.turnNumber")[10].offsetTop - document.querySelector(".floaty").scrollTop < document.querySelector(".floaty").offsetHeight) && (document.querySelectorAll("td.turnNumber")[10].offsetTop - document.querySelector(".floaty").scrollTop > 0)).toString()');
  AsyncMachine.pushOp(evaluateInPanel, 'document.querySelector("div.eventIndicator[severity=error][turn=\'20\']").dispatchEvent(new MouseEvent("mousedown"))', function(text){ return "";} );
  AsyncMachine.pushOp(evaluateInPanel, '((document.querySelectorAll("td.turnNumber")[20].offsetTop - document.querySelector(".floaty").scrollTop < document.querySelector(".floaty").offsetHeight) && (document.querySelectorAll("td.turnNumber")[20].offsetTop - document.querySelector(".floaty").scrollTop > 0)).toString()');

  AsyncMachine.runOp();
}

var testScript = test_clickMessage;
</script>
</html>
